+++ Products/CMFPlone/PloneControlPanel.py
@@ -227,7 +227,12 @@ class PloneControlPanel(PloneBaseTool, UniqueObject,
                   description='',
                   REQUEST=None,
                   ):
-        # Add an action to our list.
+        """Add an action to our list.
+
+        Note: the form in the ZMI points to this when adding an action,
+        so it must be available at the addAction url,
+        so it must have a docstring.
+        """
         if not name:
             raise ValueError('A name is required.')