+++ docs/CHANGES.rst
@@ -19,9 +19,10 @@ New features:
 
 Bug fixes:
 
-- *add item here*
+- Fix base tag differs from actual URL (fixes `#86`_).
+  [rodfersou]
 
-
+  
 4.3.11 (2016-08-15)
 -------------------
 
@@ -697,3 +698,4 @@ Fixes:
   [kleist]
 
 .. _`#1041`: https://github.com/plone/Products.CMFPlone/issues/1041
+.. _`#86`: https://github.com/plone/plone.app.layout/issues/86