+++ docs/CHANGES.rst
@@ -19,7 +19,7 @@ New features:
 
 Bug fixes:
 
-- *add item here*
+- Apply security hotfix 20160830 for ``isURLInPortal``.  [maurits]
 
 
 4.3.11 (2016-08-15)