+++ docs/CHANGES.rst
@@ -15,7 +15,8 @@ Breaking changes:
 
 New features:
 
-- *add item here*
+- Add sort_on field to search controlpanel.
+  [rodfersou]
 
 Bug fixes: